+++ prettier.config.cjs
@@ -1,6 +1,9 @@
 /* eslint-env node */
 module.exports = {
-  plugins: [require.resolve("prettier-plugin-astro")],
+  plugins: [
+    require.resolve("prettier-plugin-astro"),
+    require.resolve("prettier-plugin-svelte"),
+  ],
   overrides: [
     {
       files: "*.astro",
@@ -8,5 +11,6 @@ module.exports = {
         parser: "astro",
       },
     },
+    { files: "*.svelte", options: { parser: "svelte" } },
   ],
 };