+++ src/components/ProjectTagList.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+    export let tags: string[];
+    import '../styles/tag.css';
+</script>
+
+{#each tags as tag}
+<span class="tag">
+    #{tag}
+</span>
+{/each}